requirements.txt
Flask==3.0.2
python-dotenv==1.0.1
numpy==1.24.4
pandas==2.0.3
replicate==0.26.1
psycopg2-binary==2.9.9
Flask==3.0.2
pyyaml==6.0.1
requests==2.32.3
openai==1.30.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
sentence_transformers==2.6.0
llama-index==0.10.68
langchain==0.2.14
langchain-huggingface==0.0.3
llama-index-embeddings-langchain==0.1.2
chromadb==0.5.5
llama-index-vector-stores-chroma==0.1.10